数据类型异常

+++ genshin-map-data/genshin-map-data-core/genshin-map-data-core-model/src/main/java/site/yuanshen/data/dto/MarkerExtraDto.java
@@ -28,7 +28,7 @@ public class MarkerExtraDto {
      * 点位ID
      */
     @Schema(title = "点位ID")
-    private Integer markerId;
+    private Long markerId;
 
     /**
      * 额外特殊字段具体内容
